Added deposit script
